extract K2StatusToK2PgStatus() from storage.h into k2_util.h
