+++ src/gausskernel/storage/access/k2/storage.cpp
@@ -31,56 +31,13 @@ Copyright(c) 2022 Futurewei Cloud
 #include "catalog/pg_type.h"
 #include "fmgr/fmgr_comp.h"
 
+#include "k2_util.h"
 #include "storage.h"
 
 #include <skvhttp/dto/SKVRecord.h>
 
 namespace k2pg {
 namespace gate {
-
-int K2CodeToPGCode(int k2code) {
-    switch (k2code) {
-        case 200: // OK Codes
-        case 201:
-        case 202:
-            return ERRCODE_SUCCESSFUL_COMPLETION;
-        case 400: // Bad request
-            return ERRCODE_INTERNAL_ERROR;
-        case 403: // Forbidden, used to indicate AbortRequestTooOld in K23SI
-            return ERRCODE_SNAPSHOT_INVALID;
-        case 404: // Not found
-            return ERRCODE_SUCCESSFUL_COMPLETION;
-        case 405: // Not allowed, indicates a bug in K2 usage or operation
-        case 406: // Not acceptable, used to indicate BadFilterExpression
-        case 408: // Timeout
-            return ERRCODE_INTERNAL_ERROR;
-        case 409: // Conflict, used to indicate K23SI transaction conflicts
-            return ERRCODE_T_R_SERIALIZATION_FAILURE;
-        case 410: // Gone, indicates a partition map error
-            return ERRCODE_INTERNAL_ERROR;
-        case 412: // Precondition failed, indicates a failed K2 insert operation
-            return ERRCODE_UNIQUE_VIOLATION;
-        case 422: // Unprocessable entity, BadParameter in K23SI, indicates a bug in usage or operation
-        case 500: // Internal error, indicates a bug in K2 code
-        case 503: // Service unavailable, indicates a partition is not assigned
-        default:
-            return ERRCODE_INTERNAL_ERROR;
-    }
-
-    return ERRCODE_INTERNAL_ERROR;
-}
-
-K2PgStatus K2StatusToK2PgStatus(skv::http::Status&& status) {
-    K2PgStatus out_status{
-        .pg_code = K2CodeToPGCode(status.code),
-        .k2_code = status.code,
-        .msg = std::move(status.message),
-        .detail = ""
-    };
-
-    return out_status;
-}
-
 
 // These are types that we can push down filter operations to K2, so when we convert them we want to
 // strip out the Datum headers
@@ -128,12 +85,12 @@ K2PgStatus getSKVBuilder(K2PgOid database_oid, K2PgOid table_oid, std::shared_pt
 
     skv::http::Status catalog_status = catalog->GetCollectionNameAndSchemaName(database_oid, table_oid, collectionName, schemaName);
     if (!catalog_status.is2xxOK()) {
-        return K2StatusToK2PgStatus(std::move(catalog_status));
+        return k2pg::K2StatusToK2PgStatus(std::move(catalog_status));
     }
 
     auto [status, schema] = k2pg::TXMgr.GetSchema(collectionName, schemaName);
     if (!status.is2xxOK()) {
-        return K2StatusToK2PgStatus(std::move(status));
+        return k2pg::K2StatusToK2PgStatus(std::move(status));
     }
 
     builder = std::make_unique<skv::http::dto::SKVRecordBuilder>(collectionName, schema);
@@ -247,7 +204,7 @@ K2PgStatus makeSKVBuilderWithKeysSerialized(K2PgOid database_oid, K2PgOid table_
     std::unordered_map<int, uint32_t> attr_to_offset;
     skv::http::Status catalog_status = catalog->GetAttrNumToSKVOffset(database_oid, table_oid, attr_to_offset);
     if (!catalog_status.is2xxOK()) {
-        return K2StatusToK2PgStatus(std::move(catalog_status));
+        return k2pg::K2StatusToK2PgStatus(std::move(catalog_status));
     }
 
 
@@ -265,7 +222,7 @@ K2PgStatus makeSKVBuilderWithKeysSerialized(K2PgOid database_oid, K2PgOid table_
     uint32_t base_table_oid = 0;
     catalog_status = catalog->GetBaseTableOID(database_oid, table_oid, base_table_oid);
     if (!catalog_status.is2xxOK()) {
-        return K2StatusToK2PgStatus(std::move(catalog_status));
+        return k2pg::K2StatusToK2PgStatus(std::move(catalog_status));
     }
     uint32_t index_id = base_table_oid == table_oid ? 0 : table_oid;
 
@@ -374,13 +331,13 @@ K2PgStatus makeSKVRecordFromK2PgAttributes(K2PgOid database_oid, K2PgOid table_o
     std::unordered_map<int, uint32_t> attr_to_offset;
     skv::http::Status catalog_status = catalog->GetAttrNumToSKVOffset(database_oid, table_oid, attr_to_offset);
     if (!catalog_status.is2xxOK()) {
-        return K2StatusToK2PgStatus(std::move(catalog_status));
+        return k2pg::K2StatusToK2PgStatus(std::move(catalog_status));
     }
 
     uint32_t base_table_oid = 0;
     catalog_status = catalog->GetBaseTableOID(database_oid, table_oid, base_table_oid);
     if (!catalog_status.is2xxOK()) {
-        return K2StatusToK2PgStatus(std::move(catalog_status));
+        return k2pg::K2StatusToK2PgStatus(std::move(catalog_status));
     }
     uint32_t index_id = base_table_oid == table_oid ? 0 : table_oid;
